Party setup now allow switching position between two members in the party.
Dragging a hero from the list who is in a party into a slot occupied by another hero, will switch their places in the party. (Only works for heroes dragged from the hero list, and not from within the party frame)

Still need to implement more advanced drag and drop functions, that allows changing the order of the list elements. This will also allow to change the order in the party directly from the party frame.