Reworked the ability system
Abilities now control general information for the ability, including the damage modifiers, target information, and a list of ability effects.

The ability effects include the damage information (dice and damage roll modifiers), as well as any potential status effects. (Status effects will need a small rework as well)

This allows for a single attack roll to check whether the attack hits, followed by applying individual effects afterwards. Saving throws should be part of the effects, but that will be added alongside the status effect rework.

The hit information now contains data from the roll, so it will be possible to display the base roll and the modifiers.

Ability behaviours have been removed, since they are now part of the ability itself.